+++ src/backup/7zip/rclone.bkp.ps1
@@ -1,7 +1,3 @@
-$src = Join-Path $env:SERVICES "rclone"
-$dst = "\\raktar.local\backup\HX90\rclone"
-$arc = Join-Path $dst "rclone.$(Get-Date -Format FileDateTime).7z"
-
 if (!$env:SERVICES -Or !(Test-Path "$env:SERVICES")) {
     Write-Error "rclone.bkp.ps1 -- %SERVICES% empty or invalid. Cannot proceed"
     exit 1
@@ -11,6 +7,10 @@ if (!$env:AUTOMATION -Or !(Test-Path "$env:AUTOMATION")) {
     Write-Error "rclone.bkp.ps1 -- %AUTOMATION% empty or invalid. Cannot proceed"
     exit 1
 }
+
+$src = Join-Path $env:SERVICES "rclone"
+$dst = "\\raktar.local\backup\HX90\rclone"
+$arc = Join-Path $dst "rclone.$(Get-Date -Format FileDateTime).7z"
 
 $params = @{
     Path                    = Join-Path $env:AUTOMATION "log" "rclone.7zip.log"
@@ -22,20 +22,19 @@ Start-Transcript @params
 #-----------------------------------------------------------------------
 # Sauvegarde de la configuration de rClone sur Raktar
 Write-Host "`n`n"
-Write-Host "-----------------------------------------------------"
-Write-Host " Sauvegarde de la configuration de rClone sur Raktar "
-Write-Host "-----------------------------------------------------"
+Write-Host "----------------------------------------------"
+Write-Host " Backup of the rclone configuration on raktar "
+Write-Host "----------------------------------------------"
 
 # if the source folder doesn't exist, then there is nothing to backup
-if (!(Test-Path $src)) {
+if (!(Test-Path "$src")) {
     Write-Host "rclone.bkp.ps1 -- Source folder does not exist"
     exit 1
 }
 
-# if the destination folder does not exist, create it
-if (!(Test-Path $dst)) {
-    Write-Host "rclone.bkp.ps1 -- Creating non-existent destination"
-    New-Item $dst -ItemType Directory -Force
+if (!(Test-Path "$dst")) {
+    Write-Host "rclone.bkp.ps1 -- Destination unreachable"
+    exit 1
 }
 
 $params = @(
@@ -45,11 +44,15 @@ $params = @(
     $src
 )
 7z u @params
+if (!$?) {
+    Write-Error "rclone.bkp.ps1 -- Error while creating the archive"
+    exit 1
+}
 
 $nb_max = 5
 $bkps = Get-Item (Join-Path $dst "rclone.*.7z") | Sort-Object
 if ($bkps.Count -gt $nb_max) {
-    Write-Host "rclone.bkp.ps1 -- Enleve les fichiers de sauvegardes de trop"
+    Write-Host "rclone.bkp.ps1 -- Removing old files"
     $nb_del = $bkps.Count - $nb_max
     Remove-Item -Path ($bkps | Select-Object -First $nb_del)
 }
